# project-management-system/PMS-Swagger- Documentation.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/SriramB/ProjectManagementSystem/1.0.0
  - description: URL for PMS
    url: http://localhost:8080/api/v1
info:
  description: This API is PMS
  version: "1.0.0"
  title: PMS
  contact:
    email: suba@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
  /role/getAllRoles:
    get:
      tags:
        - developers
        - admins
        - analyst
        - manager
        - tester
      security:
        - bearerAuth: []
      summary: Search for Roles
      operationId: getAllRoles
      description: |
        Search for Roles
      responses:
        '200':
          description: List of Roles
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/RoleList'
        '400':
          description: Invalid Input
  /role/getRole/{roleId}:
    get:
      tags:
        - developers
        - admins
        - analyst
        - manager
        - tester
      security:
      - bearerAuth: []
      summary: Search for specific Role
      operationId: getRole
      description: |
        Search for specific Role
      parameters:
        - in: path
          name: roleId
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: Role
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Role'
        '400':
          description: Invalid Input
  /role/addRole:
    post:
      tags:
        - admins
        - developers
        - manager
        - tester
      security:
      - bearerAuth: []
      summary: Add a new  Role
      operationId: addRole
      description: |
        Add a new Role
      responses:
        '200':
          description: Role
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/InsertedRole'
        '400':
          description: Invalid Input
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleListContent'
  /role/updateRole/{roleId}:
    patch:
      tags:
        - admins
        - manager
      security:
      - bearerAuth: []
      summary: update existing Role
      operationId: updateRole
      description: |
        update existing Role
      parameters:
        - in: path
          name: roleId
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: Role
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/UpdatedRole'
        '400':
          description: Invalid Input
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleListContent'
  /role/deleteRole/{roleId}:
    delete:
      tags:
        - admins
        - manager
      security:
      - bearerAuth: []
      summary: delete an existing Role
      operationId: deleteRole
      description: |
        delete an existing Role
      parameters:
        - in: path
          name: roleId
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: Role
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/DeletedRole'
        '400':
          description: Invalid Input
...
  /projectService/getAllProjects:
    get:
      tags:
        - developers
        - analyst
        - manager
        - tester
        - admins
      security:
        - bearerAuth: []
      summary: fetch all the projects
      operationId: getAllProjects
      description: |
        fetch all the projects
      responses:
        '200':
          description: All Projects fetched Successfully
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/ProjectList'
        '400':
          description: Invalid Input
  /projectService/getProject/{projectId}:
    get:
      tags:
        - developers
        - analyst
        - manager
        - tester
        - admins
      security:
      - bearerAuth: []
      summary: Search for specific Project
      operationId: getProject
      description: |
        Search for specific Project
      parameters:
        - in: path
          name: projectId
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: Project Fetch Success
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/ProjectDetails'
        '400':
          description: Invalid Input
  /projectService/addProject:
    post:
      tags:
        - admins
        - manager
      security:
      - bearerAuth: []
      summary: Add a new Project
      operationId: addProject
      description: |
        Adding a new Project
      responses:
        '200':
          description: Project Successfully Added
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/AddedProject'
        '400':
          description: Invalid Input
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectListContent'
  /projectService/editProject/{projectId}:
    patch:
      tags:
        - admins
        - manager
      security:
      - bearerAuth: []
      summary: update existing Role
      operationId: updateProject
      description: |
        update existing project
      parameters:
        - in: path
          name: projectId
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: Project updation done successfully
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/UpdatedProject'
        '400':
          description: Invalid Input
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectEditContent'
  /projectService/deleteProject/{projectId}:
    delete:
      tags:
        - admins
        - manager
      security:
      - bearerAuth: []
      summary: delete an existing project
      operationId: deleteProject
      description: |
        delete an existing Project
      parameters:
        - in: path
          name: projectId
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: Project Deleted Successfully
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/DeletedProject'
        '400':
          description: Invalid Input
  /projectService/editRequirement/{requirementId}:
    patch:
      tags:
        - admins
        - manager
        - analyst
      security:
      - bearerAuth: []
      summary: update an existing project Requirement
      operationId: updateProjectRequirement
      description: |
        update an existing Project Requirement
      parameters:
        - in: path
          name: requirementId
          required: true
          schema: 
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequirementListContent'
      responses:
        '200':
          description: Edited Requirement Successfully
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/EditedRequirement'
        '400':
          description: Invalid Input
  /projectService/addRequirement/{projectId}:
    post:
      tags:
        - admins
        - manager
        - analyst
      security:
      - bearerAuth: []
      summary: add a new  Requirement to an existing project
      operationId: addProjectRequirement
      description: |
        update an existing Project Requirement
      parameters:
        - in: path
          name: projectId
          required: true
          schema: 
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequirementListContent'
      responses:
        '200':
          description: Added Requirement Successfully
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/AddedRequirement'
        '400':
          description: Invalid Input
  /projectService/deleteRequirement/{requirementId}:
    delete:
      tags:
        - admins
        - manager
        - analyst
      security:
      - bearerAuth: []
      summary: add a new  Requirement to an existing project
      operationId: deleteProjectRequirement
      description: |
        update an existing Project Requirement
      parameters:
        - in: path
          name: requirementId
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: Deleted Requirement Successfully
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/DeletedRequirement'
        '400':
          description: Invalid Input
...
components:
  securitySchemes:
    bearerAuth:   
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RoleList:
      type: object
      required:
        - Roles
      properties:
        Roles:
          items:
              $ref: '#/components/schemas/RoleListContent'
...
    RoleListContent:
      type: object
      required:
        - roleName
      properties:
        roleId:
          type: string
          example: R1
        roleName:
          type: string
          example: admin
        roleStatus:
          type: string
          example: active
        createdBy :
          type: string
          example: "U14"
        LastUpdatedBy:
          type: string
          example: "U15"
        LastUpdatedDate:
          type: object
          example: "2020-11-10T14:45:45.666+00:00"
    RoleListContentDelete:
      type: object
      required:
        - roleName
      properties:
        roleId:
          type: string
          example: R1
        roleName:
          type: string
          example: admin
        roleStatus:
          type: string
          example: InActive
        createdBy :
          type: string
          example: "U14"
        LastUpdatedBy:
          type: string
          example: "U15"
        LastUpdatedDate:
          type: object
          example: "2020-11-10T14:45:45.666+00:00"
    Role:
      type: object
      required: 
        - Role
      properties:
        Role:
          $ref: '#/components/schemas/RoleListContent'
    UpdatedRole:
      type: object
      required: 
        - Role
      properties:
        Updated Role:
          $ref: '#/components/schemas/RoleListContent'
    DeletedRole:
      type: object
      required: 
        - Role
      properties:
        Deleted:
          $ref: '#/components/schemas/RoleListContentDelete'
    InsertedRole:
      type: object
      required: 
        - Role
      properties:
        Inserted:
          $ref: '#/components/schemas/RoleListContent'
...
    ProjectList:
      type: object
      required:
        - Roles
      properties:
        Projects:
          items:
              $ref: '#/components/schemas/ProjectListContent'
    ProjectEditContent:
      properties:
          projectName:
            type: string
            example: "An Android App"
          projectDescription:
            type: string
            example: "A Mobile App to manage all users"
          endDate:
            type: object
            example: "2020-11-10T14:45:45.666+00:00"
          targetedRelease:
            type: string
            example: May
          status:
            type: string
            example: Active
...
    ProjectListContent:
      type: object
      required:
        - projectName
        - projectId
      properties:
        projectId:
          type: string
          example: P1
        projectName:
          type: string
          example: "An Android App"
        projectDescription:
          type: string
          example: "A Mobile App to manage all users"
        startDate:
          type: object
          example: "2020-11-10T14:45:45.666+00:00"
        endDate:
          type: object
          example: "2020-11-10T14:45:45.666+00:00"
        targetedRelease:
          type: string
          example: May
        status:
          type: string
          example: Active
        createdBy :
          type: string
          example: U14
        UpdatedBy:
          type: string
          example: U15
        LastUpdatedDate:
          type: object
          example: "2020-11-10T14:45:45.666+00:00"
        requirements: 
          type: array
          required:
            - requirementId
          items:
            $ref: '#/components/schemas/RequirementListContent'
    ProjectDeleteContent:
      type: object
      required:
        - projectName
        - projectId
      properties:
        projectId:
          type: string
          example: P1
        projectName:
          type: string
          example: "An Android App"
        projectDescription:
          type: string
          example: "A Mobile App to manage all users"
        startDate:
          type: object
          example: "2020-11-10T14:45:45.666+00:00"
        endDate:
          type: object
          example: "2020-11-10T14:45:45.666+00:00"
        targetedRelease:
          type: string
          example: May
        status:
          type: string
          example: InActive
        createdBy :
          type: string
          example: U14
        UpdatedBy:
          type: string
          example: U15
        LastUpdatedDate:
          type: object
          example: "2020-11-10T14:45:45.666+00:00"
        requirements: 
          type: array
          required:
            - requirementId
          items:
            $ref: '#/components/schemas/RequirementDeleteContent'
    RequirementDeleteContent:
      type: object
      required:
        - requirementName
        - requirementId
      properties:
        requirementId:
            type: string
            example: P1R1
        requirementName:
          type: string
          example: "Need Bot Support"
        requiremrntDescription:
          type: string
          example: InActive
    RequirementListContent:
      type: object
      required:
        - requirementName
        - requirementId
      properties:
        requirementId:
            type: string
            example: P1R1
        requirementName:
          type: string
          example: "Need Bot Support"
        requiremrntDescription:
          type: string
          example: Active
    ProjectDetails:
      type: object
      required: 
        - Role
      properties:
        Project Details:
          $ref: '#/components/schemas/ProjectListContent'
    AddedProject:
      type: object
      required: 
        - Role
      properties:
        Added Project:
          $ref: '#/components/schemas/ProjectListContent'
    UpdatedProject:
      type: object
      required: 
        - Role
      properties:
        Updated Project:
          $ref: '#/components/schemas/ProjectListContent'
    DeletedProject:
      type: object
      required: 
        - Role
      properties:
        Deleted Project:
          $ref: '#/components/schemas/ProjectDeleteContent'
    EditedRequirement:
      type: object
      required: 
        - Role
      properties:
        Edited Requirement:
          $ref: '#/components/schemas/ProjectListContent'
    AddedRequirement:
      type: object
      required: 
        - Role
      properties:
        Added Requirement:
          $ref: '#/components/schemas/ProjectListContent'
    DeletedRequirement:
      type: object
      required: 
        - Role
      properties:
        Deleted Requirement:
          $ref: '#/components/schemas/ProjectListContent'
...